Add button to create new todo item
We read from the existing todo items to determine what the next colour and data-name should be.

 - The colour is taken from the next colour along from the final todo. I.E. red is proceeded by yellow
 - The data-name is taken from the first available name. We test this by querying each possible data-name on the DOM

Both of these features are built with the expection of being able to delete todos in the future. Deleting would remove the entry from the DOM, and clear down it's localStorage data, allowing it's data-name to be reassigned.