Load todos dynamically from localStorage
All of the todos on the page are now generated dynamically from localStorage. This means the format for localStorage has changed to include the color as well as content for each todo.

We also save an extra item to localStorage called "order" that gets updated every time the order is changed or a new todo is added. This is read from on each page refresh to ensure todos are dynamically generated in the correct order.

The choice of next available color when adding a new todo is now based on the number of unique todos on the page. It didn't make sense to add another orange box if there aren't any green ones yet, even if the last item is red.