Add buttons to change color and delete todo
This one is a little messy in it's current implementation. We move the 'controls' element around the DOM using Javascript, but choose to reveal it based in CSS hover states. The messy part is where we need to hide the controls using Javascript when the user clicks on a todo. We don't want the controls getting in the way while they're trying to type.

To facilitate this JS/CSS hover technique, a new container has been added to wrap around the exiting todo div. This allows us to move the controls to the current todo upon hovering, without that data being stored in localStorage of the todo.

Another solution was to keep the controls in the same location on the DOM, but move it's top and left style attributes to place it in the top right corner of each todo. This however would mean re-creating the CSS hover effect using only Javascript, adding more eventListeners than we currently need.