Implement remove todo functionality
Clicking the cross on each todo will now delete it from the DOM, the localStorage, and the 'order' variable. Currently, that means the todo will be gone forever. In the future, we could implement a buffer to give you a chance to undo your deletion.

Important thing to note is we need to reset the controls back to their original location (aka not inside a todoContainer) so they don't disappear with the todo!