Polish up createExistingTodos logic
Previously this function was built around the idea of iterating through the whole localStorage variable and creating todos for each index. This meant we needed to exclude the order variable. This is no longer needed as we are now reading directly from the order variable length, and pulling out todo data based on name, rather than index.

The additional logic was technically correct at one point in time, but is now unecessary. We were checking for a condition that cannot happen.