Fix bugs identified through spec tests
Dragging and dropping was broken after the switch to "strict" Javascript, as we hadn't defined our temporary `dragIndex` and `dropIndex` variables.

Loading the site for the first time gave a blank todo because we now read the color the `.new-todo` element, but haven't set it's dataset with the `decideNextTodoColor()` function. That order has been fixed.

The mobile title size error wasn't identified through spec tests, but this is a good place to fix it.